+++ src/nomad_json_parser/example_uploads/json_mapper/example_data.json
@@ -3,56 +3,32 @@
     "title": "Create NaCl solution",
     "What_did_I_do": "Solution by stirring",
     "Steplist":[    {
-            "Steps": {
                 "step_name": "NaCl into H2O",
                 "step_description": "Put NaCl into H2O",
                 "time_unit": "seconds"
-            }
         },
         {
-            "Steps": {
                 "step_name": "Stirring 1",
                 "step_description": "Stir clockwise",
                 "How_long_did_it_take": 300.0,
                 "time_unit": "seconds"
-            }
         },
         {
-            "Steps": {
                 "step_name": "Stirring 2",
                 "step_description": "Stir anti-clockwise",
                 "How_long_did_it_take": 5.0,
                 "time_unit": "minutes"
-            }
         },
         {
-            "Steps": {
                 "step_name": "Stirring 3",
                 "step_description": "Stir clockwise",
                 "How_long_did_it_take": 600.0,
                 "time_unit": "seconds"
-            }
         }
     ],
     "Chem": {
-    "Chemicals1": {
         "name": "Rocksalt",
         "labid": "JN123_NaCl",
         "substance_formula": "NaCl"
-    },
-    "Chemicals2": {
-        "name": "Impure_Rocksalt",
-        "labid": "JN234_NaCl",
-        "substance_formula": "NaCl2"
-    },
-    "Chemicals3": {
-        "name": "NaF",
-        "labid": "JN234_NaF",
-        "substance_formula": "NaF"
-    },
-    "Chemicals4": {
-        "labid": "JN234_NaF",
-        "substance_formula": "NaF2"
     }
-}
 }
