+++ src/nomad_json_parser/parsers/__init__.py
@@ -1,14 +1,12 @@
 from nomad.config.models.plugins import (
-    ParserEntryPoint, 
+    ParserEntryPoint,
 )
-
-
 from pydantic import Field
 
 
 class JsonMapperParserEntryPoint(ParserEntryPoint):
     def load(self):
-        from nomad_json_parser.parsers.parser import JsonMapperParser
+        from nomad_json_parser.parsers.jsonmapperparser import JsonMapperParser
 
         return JsonMapperParser(**self.dict())
 
@@ -24,7 +22,7 @@ json_mapper_parser = JsonMapperParserEntryPoint(
 
 class MappedJsonParserEntryPoint(ParserEntryPoint):
     def load(self):
-        from nomad_json_parser.parsers.parser import MappedJsonParser
+        from nomad_json_parser.parsers.mappedjsonparser import MappedJsonParser
 
         return MappedJsonParser(**self.dict())
 
@@ -39,26 +37,22 @@ mapped_json_parser = MappedJsonParserEntryPoint(
 
 
 class ROCrateParserEntryPoint(ParserEntryPoint):
-
-
     json_matching_key: str = Field(
         'default_key',
         description="""
         The json mapper key to map the uploaded ROCrate object.
         """,
-    )    
-
+    )
 
     json_matching_re: str = Field(
         r'.*',
         description="""
         The regex to match to full file content.
         """,
-    )    
-
+    )
 
     def load(self):
-        from nomad_json_parser.parsers.parser import ROCrateParser
+        from nomad_json_parser.parsers.rocrateparser import ROCrateParser
 
         return ROCrateParser(**self.dict())
