+++ src/nomad_json_parser/parsers/mappedjsonparser.py
@@ -24,7 +24,6 @@ from typing import (
 import jmespath
 from nomad.datamodel import EntryArchive
 from nomad.parsing import MatchingParser
-from pyld import jsonld
 
 if TYPE_CHECKING:
     from nomad.datamodel.datamodel import (
@@ -40,7 +39,6 @@ from nomad.datamodel import ClientContext, EntryArchive
 from nomad.datamodel.metainfo.annotations import (
     Rules,
 )
-from nomad.datamodel.results import ELN, Results
 from nomad.search import (
     MetadataPagination,
     MetadataRequired,
@@ -50,229 +48,11 @@ from nomad.utils.json_transformer import Transformer
 from nomad_material_processing.utils import create_archive
 
 from nomad_json_parser.schema_packages.jsonimport import (
-    JsonMapper,
-    MainMapper,
     MappedJson,
-    MapperRule,
-    MatchKey,
-    RepeatKey,
     RepeatPath,
-    RuleCondition,
-    SubSectionMapper,
     createrulesjson,
     get_class,
 )
-
-
-def create_rules(subsection, key, logger):
-    rules = []
-    if 'rules' in subsection:
-        for rulekey in subsection['rules'].keys():
-            rule = subsection['rules'][rulekey]
-            rulesection = MapperRule()
-            try:
-                if not ('source' in rule.keys() and 'target' in rule.keys()):
-                    logger.error(
-                        f'Rule {rulekey} from Subsection {key} is '
-                        'missing source or target key.'
-                    )
-                rulesection.name = rulekey
-                rulesection.source = rule['source']
-                rulesection.target = rule['target']
-                if 'default_value' in rule.keys():
-                    rulesection.default_value = rule['default_value']
-                if 'use_rule' in rule.keys():
-                    rulesection.use_rule = rule['use_rule']
-                if 'conditions' in rule.keys():
-                    condlist = []
-                    for condition in rule['conditions']:
-                        conditionssection = RuleCondition()
-                        condname = next(iter(condition))
-                        conditionssection.name = condname
-                        conditionssection.regex_path = condition[condname]['regex_path']
-                        conditionssection.regex_pattern = condition[condname][
-                            'regex_pattern'
-                        ]
-                        condlist.append(conditionssection)
-                    rulesection.conditions = condlist
-                if 'target_type' in rule.keys():
-                    rulesection.target_type = rule['target_type']
-                if 'match_key' in rule.keys():
-                    all_keys = []
-                    for m_key in range(10):
-                        str_m_key = f'match_key{m_key}'.strip('0')
-                        if str_m_key not in rule.keys():
-                            break
-                        matchkey = MatchKey()
-                        matchkey.name = f'Match for part {m_key}.'
-                        keys_list = []
-                        for s in rule[str_m_key]:
-                            repkey = RepeatKey()
-                            repkey.name = str(s)
-                            keys_list.append(repkey)
-                        matchkey.match_keys = keys_list
-                        all_keys.append(matchkey)
-                    rulesection.match_key = all_keys
-            except AttributeError:
-                rulesection.name = f'{rulekey}_to_{rule}'
-                rulesection.source = rulekey
-                rulesection.target = rule
-            rules.append(rulesection)
-    else:
-        logger.warning(
-            f'Rules section is missing from Subsection {key}. \
-                No mapping will be done.'
-        )
-    return rules
-
-
-def create_mainmapping(subsection, key, logger, archive):  # noqa: PLR0912, PLR0915
-    sectionclass = MainMapper()
-    if (
-        'main_key' in subsection
-        or 'is_archive' in subsection
-        or 'repeats' in subsection
-        or 'repeat_paths' in subsection
-    ):
-        logger.error(
-            'Main section of json mapper should not contain \
-                main_key or is_archive or repeats or repeat_paths.'
-        )
-    sectionclass.name = key
-    try:
-        sectionclass.path_to_schema = subsection['schema']
-    except KeyError:
-        logger.error(f'schema is missing from Subsection {key}.')
-    sectionclass.rules = create_rules(subsection, key, logger)
-    sectionclass.normalize(archive, logger)
-    return sectionclass
-
-
-def create_submapping(subsection, key, logger, archive):  # noqa: PLR0912, PLR0915
-    sectionclass = SubSectionMapper()
-    try:
-        sectionclass.main_key = subsection['main_key']
-    except KeyError:
-        logger.error(f'main_key is missing from Subsection {key}.')
-    if 'is_archive' in subsection:
-        sectionclass.is_archive = subsection['is_archive']
-    if 'repeats' in subsection:
-        sectionclass.repeats = subsection['repeats']
-        if 'repeat_paths' in subsection:
-            repeat_paths = []
-            for path in subsection['repeat_paths']:
-                repeat = RepeatPath()
-                repeat.name = path
-                repeat_paths.append(repeat)
-            sectionclass.repeat_paths = repeat_paths
-            if 'repeat_keys' in subsection:
-                keys_list = []
-                for s in subsection['repeat_keys']:
-                    repkey = RepeatKey()
-                    repkey.name = str(s)
-                    keys_list.append(repkey)
-                sectionclass.repeat_keys = keys_list
-    if 'repeat_paths' in subsection and 'repeats' not in subsection:
-        logger.warning('repeat_paths found but not repeats, ignoring repeat_paths.')
-    if 'subsections' in subsection:
-        subsections = []
-        for subkey in subsection['subsections'].keys():
-            logger.info(subkey, subsection['subsections'][subkey])
-            subsectionclass = create_submapping(
-                subsection['subsections'][subkey], subkey, logger, archive
-            )  # noqa: E501
-            subsections.append(subsectionclass)
-        sectionclass.subsection_mappings = subsections
-    sectionclass.name = key
-    try:
-        sectionclass.path_to_schema = subsection['schema']
-    except KeyError:
-        logger.error(f'schema is missing from Subsection {key}.')
-    sectionclass.rules = create_rules(subsection, key, logger)
-    sectionclass.normalize(archive, logger)
-    return sectionclass
-
-
-def create_sectionclass(jsonfile, logger, archive):  # noqa: PLR0912, PLR0915
-    main_found = False
-    subsections = []
-    for key in jsonfile.keys():
-        if key in {'$json_mapper_class_key', '$json_mapper_version'}:
-            continue
-        subsection = jsonfile[key]
-        if 'is_main' in subsection and subsection['is_main'] == 'True':
-            if not main_found:
-                main_found = True
-                main_mapping = create_mainmapping(subsection, key, logger, archive)
-            else:
-                logger.error('is_main can only be in one Subsection.')
-        else:
-            sectionclass = create_submapping(subsection, key, logger, archive)
-            subsections.append(sectionclass)
-    return main_mapping, subsections
-
-
-class JsonMapperParser(MatchingParser):
-    def set_entrydata_definition(self):
-        self.entrydata_definition = JsonMapper
-
-    def parse(self, mainfile: str, archive: EntryArchive, logger) -> None:  # noqa: PLR0912, PLR0915
-        self.set_entrydata_definition()
-        data_file_with_path = mainfile.rsplit('raw/', maxsplit=1)[-1]
-        entry = self.entrydata_definition()
-        entry.mapper_file = data_file_with_path
-
-        if not archive.results:
-            archive.results = Results(eln=ELN())
-        if not archive.results.eln:
-            archive.results.eln = ELN()
-        archive.results.eln.sections = ['JsonMapper']
-
-        if entry.mapper_file:
-            with archive.m_context.raw_file(entry.mapper_file, 'r') as file:
-                jsonfile = json.load(file)
-
-            try:
-                entry.mapper_key = jsonfile['$json_mapper_class_key']
-                archive.results.eln.lab_ids = [entry.mapper_key]
-                if '$json_mapper_version' in jsonfile.keys():
-                    entry.mapper_version = jsonfile['$json_mapper_version']
-                else:
-                    entry.mapper_version = 1
-                archive.results.eln.tags = [entry.mapper_version]
-            except KeyError:
-                logger.error(
-                    'Missing keys for jsonmapper file ($json_mapper_class_key).'
-                )
-            logger.info(
-                'Starting search for already existing mappers with\
-                      same key and version.'
-            )
-            if not isinstance(archive.m_context, ClientContext):
-                search_result = search(
-                    owner='all',
-                    query={
-                        'data.mapper_key#nomad_json_parser.schema_packages.jsonimport.JsonMapper': entry.mapper_key,  # noqa: E501
-                        'data.mapper_version#nomad_json_parser.schema_packages.jsonimport.JsonMapper': entry.mapper_version,  # noqa: E501
-                    },
-                    user_id=archive.metadata.main_author.user_id,
-                )
-                if len(search_result.data) > 1:
-                    logger.error(
-                        'At least one mapper with the same key and\
-                              version has been found.'
-                    )
-
-            entry.main_mapping, entry.subsection_mappings = create_sectionclass(
-                jsonfile, logger, archive
-            )
-            if entry.main_mapping is None:
-                logger.error('No main mapping found.')
-
-        archive.data = entry
-        archive.metadata.entry_name = (
-            f'JsonMapper_{entry.mapper_key}_v{entry.mapper_version}'
-        )
 
 
 def correcttypes(transformed_sub, subclass_rules):
@@ -291,7 +71,7 @@ def correcttypes(transformed_sub, subclass_rules):
     return transformed_sub
 
 
-def transform_subclass(  # noqa: PLR0913
+def transform_subclass(  # noqa: PLR0913, PLR0917
     subclass_mapping,
     logger,
     jsonfile,
@@ -351,7 +131,7 @@ def transform_subclass(  # noqa: PLR0913
                 sub_ref = create_archive(
                     subsubclass,
                     archive,
-                    f'{data_file_with_path.rsplit("/", maxsplit=1)[0]}/{subsubclass.name}.archive.json',
+                    f'{subsubclass.name}.archive.json',
                 )
                 archive_list.append(sub_ref)
                 setattr(subclass, subsectionmap['main_key'], sub_ref)
@@ -482,7 +262,7 @@ def checkforvalidrulematch(backjson, matchrule):
     return True
 
 
-def ruleblock_resolve(ruleblock, jsonfile, logger):
+def ruleblock_resolve(ruleblock, jsonfile, logger):  # noqa: PLR0912
     newruleblock = []
     for rule in ruleblock:
         newrule = deepcopy(rule)
@@ -589,7 +369,7 @@ def expand_mapper(mapper):
     return mapper
 
 
-def map_with_nesting(
+def map_with_nesting(  # noqa: PLR0912, PLR0913, PLR0917
     mapper, mapname, logger, archive, jsonfile, archive_list, data_file_with_path
 ):  # noqa: PLR0912, PLR0913
     mapkey = ''
@@ -650,7 +430,7 @@ def map_with_nesting(
                     sub_ref = create_archive(
                         subsubclass,
                         archive,
-                        f'{data_file_with_path.rsplit("/", maxsplit=1)[0]}/{subsubclass.name}.archive.json',
+                        f'{subsubclass.name}.archive.json',  # noqa: E501
                     )
                     archive_list.append(sub_ref)
                     setattr(subclass, shortened_mainkey, sub_ref)
@@ -671,6 +451,8 @@ class MappedJsonParser(MatchingParser):
         data_file_with_path = mainfile.rsplit('raw/', maxsplit=1)[-1]
         entry = self.entrydata_definition()
         entry.json_file = data_file_with_path
+
+        logger.error(mainfile)
 
         if entry.json_file:
             with archive.m_context.raw_file(entry.json_file, 'r') as file:
@@ -762,7 +544,7 @@ class MappedJsonParser(MatchingParser):
             main_ref = create_archive(
                 mainclass,
                 archive,
-                f'{data_file_with_path.rsplit("/", maxsplit=1)[0]}/{mainclass.name}.archive.json',
+                f'{mainclass.name}.archive.json',
             )
             archive_list.append(main_ref)
             archive_list.reverse()  # put main entry as first
@@ -772,115 +554,3 @@ class MappedJsonParser(MatchingParser):
         archive.metadata.entry_name = (
             f'{data_file}_MappedJson_{entry.mapper_key}_v{entry.mapper_version}'
         )
-
-
-class ROCrateParser(MatchingParser):
-    def __init__(
-        self,
-        json_matching_key: str | None = None,
-        json_matching_re: str = r'.*',
-        **kwargs,
-    ) -> None:
-        super().__init__(**kwargs)
-        self.json_matching_key = json_matching_key
-        self.json_matching_re = re.compile(json_matching_re)
-
-    def is_mainfile(
-        self,
-        filename: str,
-        mime: str,
-        buffer: bytes,
-        decoded_buffer: str,
-        compression: str | None = None,
-    ) -> bool:
-        # First do standard matching
-        if not super().is_mainfile(filename, mime, buffer, decoded_buffer, compression):
-            return False
-
-        # Additional custom logic to ensure this is the ONLY parser that matches
-        # For example, check that this is specifically a PLD file
-
-        # Read the file to make sure it's really a PLD file
-        with open(filename) as f:
-            content = f.read()
-        # Only proceed if it's definitely a PLD file
-        if self.json_matching_re.search(content) is None:
-            return False
-
-        return True
-
-    def parse(self, mainfile: str, archive: EntryArchive, logger) -> None:  # noqa: PLR0912, PLR0915
-        data_file_with_path = mainfile.rsplit('raw/', maxsplit=1)[-1]
-
-        attrs = vars(self)
-        for key, value in attrs.items():
-            logger.error(f'{key}: {value}')
-
-        with archive.m_context.raw_file(data_file_with_path, 'r') as file:
-            data = json.load(file)
-
-        # fix file pathes (TODO: local pathes would fix the issue)
-        for j in range(len(data['@graph'])):
-            if data['@graph'][j]['@type'] == 'File':
-                data = json.loads(
-                    json.dumps(data).replace(
-                        data['@graph'][j]['@id'],
-                        data['@graph'][j]['@id'].replace(
-                            './',
-                            './' + data_file_with_path.strip('ro-crate-metadata.json'),
-                        ),
-                    )
-                )
-
-        # fix base id
-        expandeddata = jsonld.expand(
-            data,
-            options={'base': 'file://'},
-            # this will replace "./" with "file:///"
-        )
-
-        frame = {
-            '@context': [
-                'https://w3id.org/ro/crate/1.2/context',
-                {
-                    '@base': 'http://example.org/base/',
-                    '@vocab': 'http://example.org/base/',
-                },
-            ],
-            '@id': 'file:./',
-            '@embed': '@always',
-        }
-
-        frameddata = jsonld.frame(expandeddata, frame)
-
-        compact_context = {
-            '@context': [
-                'https://w3id.org/ro/crate/1.2/context',
-                {
-                    '@base': 'file:./',  # will replace "file:///" with "./"
-                    'type': '@type',  # optional for clean keywords
-                    'id': '@id',  # optional for clean keywords
-                    'base': '@base',
-                },
-            ]
-        }
-
-        compacteddata = jsonld.compact(
-            frameddata, compact_context
-        )  # save the framed data to file
-
-        compacteddatawithkey = {
-            '$mapped_json_class_key': self.json_matching_key,
-            **compacteddata,
-        }
-
-        filename = data_file_with_path.replace(
-            'ro-crate-metadata.json', 'frameddata.json'
-        )
-
-        with archive.m_context.raw_file(filename, 'w') as outfile:
-            json.dump(compacteddatawithkey, outfile)
-
-        if self.json_matching_key != 'default_key':
-            toparse = MappedJsonParser(json_file=filename)
-            toparse.parse(filename, archive, logger)
